+++ hubdle-gradle-plugin/main/kotlin/com/javiersc/hubdle/extensions/kotlin/multiplatform/targets/HubdleKotlinMultiplatformIOSExtension.kt
@@ -31,7 +31,7 @@ constructor(
     override val project: Project
         get() = super.project
 
-    override val isEnabled: Property<Boolean> = property { true }
+    override val isEnabled: Property<Boolean> = property { false }
 
     override val priority: Priority = Priority.P3
 
@@ -90,7 +90,7 @@ constructor(
                 iosX64()
             }
         }
-        configurable(priority = Priority.P4) {
+        configurable(priority = Priority.P6) {
             configure<KotlinMultiplatformExtension> {
                 val commonMain: KotlinSourceSet by sourceSets.getting
                 val iosArm32Main: KotlinSourceSet? = sourceSets.findByName("iosArm32Main")